cell size reduction upon oxidative stress
Thomas Kehl-Fie (committee member for Josh) mentioned in a recent
meeting that bacterial cells will experience a reduction in cell size
upon hydrogen peroxide treatment, which could distort fluorescence
measurements by flow. I tested this with Hanxi's data and it seems to
apply to yeast cells as well, with the exception of the 10 mM H2O2
treated sample, which showed a right-shift (getting bigger) in cell size
compared with the mock sample.